# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Core API Docs",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/admin/categories": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get category list",
                "tags": [
                    "Category"
                ],
                "summary": "Get category list",
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BaseOutput"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.CategoryDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create category",
                "tags": [
                    "Category"
                ],
                "summary": "Create category",
                "parameters": [
                    {
                        "description": "Register",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.StoreCategoryDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/categories/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update category",
                "tags": [
                    "Category"
                ],
                "summary": "Update category",
                "parameters": [
                    {
                        "description": "Register",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.StoreCategoryDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "CategoryID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/media": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get media list",
                "tags": [
                    "Media"
                ],
                "summary": "Get media list",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "userID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BaseOutput"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.MediaDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Upload media",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Upload media",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Upload",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/admin/roles": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get role list",
                "tags": [
                    "Role"
                ],
                "summary": "Get role list",
                "responses": {
                    "200": {
                        "description": "Role",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BaseOutput"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.RoleDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create role",
                "tags": [
                    "Role"
                ],
                "summary": "Create role",
                "parameters": [
                    {
                        "description": "Role",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateRoleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/roles/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update role",
                "tags": [
                    "Role"
                ],
                "summary": "Update role",
                "parameters": [
                    {
                        "description": "Role",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateRoleDTO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "RoleID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseOutput"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Login system",
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BaseOutput"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Register system",
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Register",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseOutput"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.BaseOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/types.PaginationDTO"
                }
            }
        },
        "types.BaseRelationDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.CategoryDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "$ref": "#/definitions/types.BaseRelationDTO"
                },
                "parentId": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "types.CreateRoleDTO": {
            "type": "object",
            "required": [
                "name",
                "permissions"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "types.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.MediaDTO": {
            "type": "object",
            "properties": {
                "fileName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "types.PaginationDTO": {
            "type": "object",
            "properties": {
                "numberOfElements": {
                    "type": "integer"
                },
                "totalElements": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "types.RegisterRequest": {
            "type": "object",
            "required": [
                "full_name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.RoleDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.StoreCategoryDTO": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}